Fix #61: Refactor model.rs and management.rs
