+++ libwebauthn/src/management.rs
@@ -1,548 +1,5 @@
-use serde_cbor::ser::to_vec;
-use std::time::Duration;
-use tracing::info;
+mod bio_enrollment;
+pub use bio_enrollment::BioEnrollment;
 
-use crate::pin::{PinProvider, PinUvAuthProtocol};
-use crate::proto::ctap2::{
-    Ctap2AuthenticatorConfigCommand, Ctap2BioEnrollmentFingerprintKind, Ctap2BioEnrollmentRequest,
-    Ctap2BioEnrollmentTemplateId, Ctap2GetInfoResponse, Ctap2LastEnrollmentSampleStatus,
-};
-pub use crate::transport::error::{CtapError, Error, TransportError};
-use crate::transport::Channel;
-use crate::webauthn::handle_errors;
-use crate::webauthn::{user_verification, UsedPinUvAuthToken};
-use crate::{
-    ops::webauthn::UserVerificationRequirement,
-    proto::ctap2::{
-        Ctap2, Ctap2AuthTokenPermissionRole, Ctap2AuthenticatorConfigRequest,
-        Ctap2BioEnrollmentModality, Ctap2UserVerifiableRequest,
-    },
-};
-use async_trait::async_trait;
-use serde_bytes::ByteBuf;
-
-#[async_trait]
-pub trait AuthenticatorConfig {
-    async fn toggle_always_uv(
-        &mut self,
-        pin_provider: &Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error>;
-
-    async fn enable_enterprise_attestation(
-        &mut self,
-        pin_provider: &Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error>;
-
-    async fn set_min_pin_length(
-        &mut self,
-        new_pin_length: u64,
-        pin_provider: &Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error>;
-
-    async fn force_change_pin(
-        &mut self,
-        force: bool,
-        pin_provider: &Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error>;
-
-    async fn set_min_pin_length_rpids(
-        &mut self,
-        rpids: Vec<String>,
-        pin_provider: &Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error>;
-}
-
-#[async_trait]
-impl<C> AuthenticatorConfig for C
-where
-    C: Channel,
-{
-    async fn toggle_always_uv(
-        &mut self,
-        pin_provider: &Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error> {
-        let mut req = Ctap2AuthenticatorConfigRequest::new_toggle_always_uv();
-
-        loop {
-            let uv_auth_used = user_verification(
-                self,
-                UserVerificationRequirement::Required,
-                &mut req,
-                pin_provider,
-                timeout,
-            )
-            .await?;
-            // On success, this is an all-empty Ctap2AuthenticatorConfigResponse
-            handle_errors!(
-                self,
-                self.ctap2_authenticator_config(&req, timeout).await,
-                uv_auth_used
-            )
-        }
-    }
-
-    async fn enable_enterprise_attestation(
-        &mut self,
-        pin_provider: &Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error> {
-        let mut req = Ctap2AuthenticatorConfigRequest::new_enable_enterprise_attestation();
-
-        loop {
-            let uv_auth_used = user_verification(
-                self,
-                UserVerificationRequirement::Required,
-                &mut req,
-                pin_provider,
-                timeout,
-            )
-            .await?;
-            // On success, this is an all-empty Ctap2AuthenticatorConfigResponse
-            handle_errors!(
-                self,
-                self.ctap2_authenticator_config(&req, timeout).await,
-                uv_auth_used
-            )
-        }
-    }
-
-    async fn set_min_pin_length(
-        &mut self,
-        new_pin_length: u64,
-        pin_provider: &Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error> {
-        let mut req = Ctap2AuthenticatorConfigRequest::new_set_min_pin_length(new_pin_length);
-
-        loop {
-            let uv_auth_used = user_verification(
-                self,
-                UserVerificationRequirement::Required,
-                &mut req,
-                pin_provider,
-                timeout,
-            )
-            .await?;
-            // On success, this is an all-empty Ctap2AuthenticatorConfigResponse
-            handle_errors!(
-                self,
-                self.ctap2_authenticator_config(&req, timeout).await,
-                uv_auth_used
-            )
-        }
-    }
-
-    async fn force_change_pin(
-        &mut self,
-        force: bool,
-        pin_provider: &Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error> {
-        let mut req = Ctap2AuthenticatorConfigRequest::new_force_change_pin(force);
-
-        loop {
-            let uv_auth_used = user_verification(
-                self,
-                UserVerificationRequirement::Required,
-                &mut req,
-                pin_provider,
-                timeout,
-            )
-            .await?;
-            // On success, this is an all-empty Ctap2AuthenticatorConfigResponse
-            handle_errors!(
-                self,
-                self.ctap2_authenticator_config(&req, timeout).await,
-                uv_auth_used
-            )
-        }
-    }
-
-    async fn set_min_pin_length_rpids(
-        &mut self,
-        rpids: Vec<String>,
-        pin_provider: &Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error> {
-        let mut req = Ctap2AuthenticatorConfigRequest::new_set_min_pin_length_rpids(rpids);
-        loop {
-            let uv_auth_used = user_verification(
-                self,
-                UserVerificationRequirement::Required,
-                &mut req,
-                pin_provider,
-                timeout,
-            )
-            .await?;
-            // On success, this is an all-empty Ctap2AuthenticatorConfigResponse
-            handle_errors!(
-                self,
-                self.ctap2_authenticator_config(&req, timeout).await,
-                uv_auth_used
-            )
-        }
-    }
-}
-
-impl Ctap2UserVerifiableRequest for Ctap2AuthenticatorConfigRequest {
-    fn ensure_uv_set(&mut self) {
-        // No-op
-    }
-
-    fn calculate_and_set_uv_auth(
-        &mut self,
-        uv_proto: &Box<dyn PinUvAuthProtocol>,
-        uv_auth_token: &[u8],
-    ) {
-        // pinUvAuthParam (0x04): the result of calling
-        // authenticate(pinUvAuthToken, 32×0xff || 0x0d || uint8(subCommand) || subCommandParams).
-        let mut data = vec![0xff; 32];
-        data.push(0x0D);
-        data.push(self.subcommand as u8);
-        if self.subcommand == Ctap2AuthenticatorConfigCommand::SetMinPINLength {
-            data.extend(to_vec(&self.subcommand_params).unwrap());
-        }
-        let uv_auth_param = uv_proto.authenticate(uv_auth_token, &data);
-        self.protocol = Some(uv_proto.version());
-        self.uv_auth_param = Some(ByteBuf::from(uv_auth_param));
-    }
-
-    fn client_data_hash(&self) -> &[u8] {
-        unreachable!()
-    }
-
-    fn permissions(&self) -> Ctap2AuthTokenPermissionRole {
-        return Ctap2AuthTokenPermissionRole::AUTHENTICATOR_CONFIGURATION;
-    }
-
-    fn permissions_rpid(&self) -> Option<&str> {
-        None
-    }
-
-    fn can_use_uv(&self, info: &Ctap2GetInfoResponse) -> bool {
-        info.option_enabled("uvAcfg")
-    }
-}
-
-#[async_trait]
-pub trait BioEnrollment {
-    async fn get_bio_modality(
-        &mut self,
-        timeout: Duration,
-    ) -> Result<Ctap2BioEnrollmentModality, Error>;
-    async fn get_fingerprint_sensor_info(
-        &mut self,
-        timeout: Duration,
-    ) -> Result<Ctap2BioEnrollmentFingerprintSensorInfo, Error>;
-    async fn get_bio_enrollments(
-        &mut self,
-        pin_provider: &mut Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<Vec<Ctap2BioEnrollmentTemplateId>, Error>;
-    async fn remove_bio_enrollment(
-        &mut self,
-        template_id: &[u8],
-        pin_provider: &mut Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error>;
-    async fn rename_bio_enrollment(
-        &mut self,
-        template_id: &[u8],
-        template_friendly_name: &str,
-        pin_provider: &mut Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error>;
-    async fn start_new_bio_enrollment(
-        &mut self,
-        pin_provider: &mut Box<dyn PinProvider>,
-        enrollment_timeout: Option<Duration>,
-        timeout: Duration,
-    ) -> Result<(Vec<u8>, Ctap2LastEnrollmentSampleStatus, u64), Error>;
-    async fn capture_next_bio_enrollment_sample(
-        &mut self,
-        template_id: &[u8],
-        pin_provider: &mut Box<dyn PinProvider>,
-        enrollment_timeout: Option<Duration>,
-        timeout: Duration,
-    ) -> Result<(Ctap2LastEnrollmentSampleStatus, u64), Error>;
-    async fn cancel_current_bio_enrollment(
-        &mut self,
-        pin_provider: &mut Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error>;
-}
-
-#[derive(Debug, Clone)]
-pub struct Ctap2BioEnrollmentFingerprintSensorInfo {
-    pub fingerprint_kind: Ctap2BioEnrollmentFingerprintKind,
-    pub max_capture_samples_required_for_enroll: Option<u64>,
-    pub max_template_friendly_name: Option<u64>,
-}
-
-#[async_trait]
-impl<C> BioEnrollment for C
-where
-    C: Channel,
-{
-    async fn get_bio_modality(
-        &mut self,
-        timeout: Duration,
-    ) -> Result<Ctap2BioEnrollmentModality, Error> {
-        let req = Ctap2BioEnrollmentRequest::new_get_modality();
-        // No UV needed
-        let resp = self.ctap2_bio_enrollment(&req, timeout).await?;
-        resp.modality.ok_or(Error::Ctap(CtapError::Other))
-    }
-
-    async fn get_fingerprint_sensor_info(
-        &mut self,
-        timeout: Duration,
-    ) -> Result<Ctap2BioEnrollmentFingerprintSensorInfo, Error> {
-        let req = Ctap2BioEnrollmentRequest::new_fingerprint_sensor_info();
-        // No UV needed
-        let resp = self.ctap2_bio_enrollment(&req, timeout).await?;
-        Ok(Ctap2BioEnrollmentFingerprintSensorInfo {
-            fingerprint_kind: resp.fingerprint_kind.ok_or(Error::Ctap(CtapError::Other))?,
-            max_capture_samples_required_for_enroll: resp
-                .max_capture_samples_required_for_enroll
-                .clone(),
-            max_template_friendly_name: resp.max_template_friendly_name,
-        })
-    }
-
-    async fn get_bio_enrollments(
-        &mut self,
-        pin_provider: &mut Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<Vec<Ctap2BioEnrollmentTemplateId>, Error> {
-        let mut req = Ctap2BioEnrollmentRequest::new_enumerate_enrollments();
-
-        let resp = loop {
-            let uv_auth_used = user_verification(
-                self,
-                UserVerificationRequirement::Preferred,
-                &mut req,
-                pin_provider,
-                timeout,
-            )
-            .await?;
-
-            // On success, this is an all-empty Ctap2AuthenticatorConfigResponse
-            handle_errors!(
-                self,
-                self.ctap2_bio_enrollment(&req, timeout).await,
-                uv_auth_used
-            )
-        };
-        Ok(resp?.template_infos.unwrap_or_default())
-    }
-
-    async fn remove_bio_enrollment(
-        &mut self,
-        template_id: &[u8],
-        pin_provider: &mut Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error> {
-        let mut req = Ctap2BioEnrollmentRequest::new_remove_enrollment(template_id);
-
-        loop {
-            let uv_auth_used = user_verification(
-                self,
-                UserVerificationRequirement::Preferred,
-                &mut req,
-                pin_provider,
-                timeout,
-            )
-            .await?;
-
-            // On success, this is an all-empty Ctap2AuthenticatorConfigResponse
-            handle_errors!(
-                self,
-                self.ctap2_bio_enrollment(&req, timeout).await,
-                uv_auth_used
-            )
-        }?;
-
-        // "If there is an exiting enrollment with passed in templateInfo, delete that enrollment and return CTAP2_OK."
-        // So, the resulting Response will be empty on success.
-        Ok(())
-    }
-
-    async fn rename_bio_enrollment(
-        &mut self,
-        template_id: &[u8],
-        template_friendly_name: &str,
-        pin_provider: &mut Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error> {
-        let mut req =
-            Ctap2BioEnrollmentRequest::new_rename_enrollment(template_id, template_friendly_name);
-        loop {
-            let uv_auth_used = user_verification(
-                self,
-                UserVerificationRequirement::Preferred,
-                &mut req,
-                pin_provider,
-                timeout,
-            )
-            .await?;
-
-            // On success, this is an all-empty Ctap2AuthenticatorConfigResponse
-            handle_errors!(
-                self,
-                self.ctap2_bio_enrollment(&req, timeout).await,
-                uv_auth_used
-            )
-        }?;
-        // "If there is an exiting enrollment with passed in templateInfo, delete that enrollment and return CTAP2_OK."
-        // So, the resulting Response will be empty on success.
-        Ok(())
-    }
-
-    async fn start_new_bio_enrollment(
-        &mut self,
-        pin_provider: &mut Box<dyn PinProvider>,
-        enrollment_timeout: Option<Duration>,
-        timeout: Duration,
-    ) -> Result<(Vec<u8>, Ctap2LastEnrollmentSampleStatus, u64), Error> {
-        let mut req = Ctap2BioEnrollmentRequest::new_start_new_enrollment(enrollment_timeout);
-
-        let resp = loop {
-            let uv_auth_used = user_verification(
-                self,
-                UserVerificationRequirement::Preferred,
-                &mut req,
-                pin_provider,
-                timeout,
-            )
-            .await?;
-
-            // On success, this is an all-empty Ctap2AuthenticatorConfigResponse
-            handle_errors!(
-                self,
-                self.ctap2_bio_enrollment(&req, timeout).await,
-                uv_auth_used
-            )
-        }?;
-
-        // TODO: authenticators MUST return these, but should we guard against devices that don't, instead of unwrapping here?
-        let remaining_samples = resp.remaining_samples.unwrap();
-        let template_id = resp.template_id.unwrap().clone();
-        let sample_status = resp.last_enroll_sample_status.unwrap();
-        Ok((template_id.to_vec(), sample_status, remaining_samples))
-    }
-
-    async fn capture_next_bio_enrollment_sample(
-        &mut self,
-        template_id: &[u8],
-        pin_provider: &mut Box<dyn PinProvider>,
-        enrollment_timeout: Option<Duration>,
-        timeout: Duration,
-    ) -> Result<(Ctap2LastEnrollmentSampleStatus, u64), Error> {
-        let mut req =
-            Ctap2BioEnrollmentRequest::new_next_enrollment(template_id, enrollment_timeout);
-
-        let resp = loop {
-            let uv_auth_used = user_verification(
-                self,
-                UserVerificationRequirement::Preferred,
-                &mut req,
-                pin_provider,
-                timeout,
-            )
-            .await?;
-
-            // On success, this is an all-empty Ctap2AuthenticatorConfigResponse
-            handle_errors!(
-                self,
-                self.ctap2_bio_enrollment(&req, timeout).await,
-                uv_auth_used
-            )
-        }?;
-
-        // TODO: authenticators MUST return these, but should we guard against devices that don't, instead of unwrapping here?
-        let remaining_samples = resp.remaining_samples.unwrap();
-        let sample_status = resp.last_enroll_sample_status.unwrap();
-        Ok((sample_status, remaining_samples))
-    }
-
-    async fn cancel_current_bio_enrollment(
-        &mut self,
-        pin_provider: &mut Box<dyn PinProvider>,
-        timeout: Duration,
-    ) -> Result<(), Error> {
-        let mut req = Ctap2BioEnrollmentRequest::new_cancel_current_enrollment();
-
-        loop {
-            let uv_auth_used = user_verification(
-                self,
-                UserVerificationRequirement::Preferred,
-                &mut req,
-                pin_provider,
-                timeout,
-            )
-            .await?;
-
-            // On success, this is an all-empty Ctap2AuthenticatorConfigResponse
-            handle_errors!(
-                self,
-                self.ctap2_bio_enrollment(&req, timeout).await,
-                uv_auth_used
-            )
-        }?;
-
-        // "Authenticator on receiving such command, cancels current ongoing enrollment, if any, and returns CTAP2_OK."
-        // So, the resulting Response will be empty on success.
-        Ok(())
-    }
-}
-
-impl Ctap2UserVerifiableRequest for Ctap2BioEnrollmentRequest {
-    fn ensure_uv_set(&mut self) {
-        // No-op
-    }
-
-    fn calculate_and_set_uv_auth(
-        &mut self,
-        uv_proto: &Box<dyn PinUvAuthProtocol>,
-        uv_auth_token: &[u8],
-    ) {
-        // pinUvAuthParam (0x05): authenticate(pinUvAuthToken, fingerprint (0x01) || enumerateEnrollments (0x04)).
-        let mut data = match self.subcommand {
-            None => unreachable!(),
-            Some(x) => {
-                let data = vec![Ctap2BioEnrollmentModality::Fingerprint as u8, x as u8];
-                data
-            }
-        };
-        // e.g. "Authenticator calls verify(pinUvAuthToken, fingerprint (0x01) || removeEnrollment (0x06) || subCommandParams, pinUvAuthParam)"
-        if let Some(params) = &self.subcommand_params {
-            data.extend(to_vec(params).unwrap());
-        }
-        let uv_auth_param = uv_proto.authenticate(uv_auth_token, &data);
-        self.protocol = Some(uv_proto.version());
-        self.uv_auth_param = Some(ByteBuf::from(uv_auth_param));
-    }
-
-    fn client_data_hash(&self) -> &[u8] {
-        unreachable!()
-    }
-
-    fn permissions(&self) -> Ctap2AuthTokenPermissionRole {
-        return Ctap2AuthTokenPermissionRole::BIO_ENROLLMENT;
-    }
-
-    fn permissions_rpid(&self) -> Option<&str> {
-        None
-    }
-
-    fn can_use_uv(&self, info: &Ctap2GetInfoResponse) -> bool {
-        info.option_enabled("uvBioEnroll")
-    }
-}
+mod authenticator_config;
+pub use authenticator_config::AuthenticatorConfig;
