+++ libwebauthn/src/proto/ctap2/model.rs
@@ -1,26 +1,36 @@
-use std::collections::BTreeMap;
-use std::collections::HashMap;
-use std::convert::TryFrom;
-use std::io::Cursor as IOCursor;
-use std::time::Duration;
-
-use byteorder::{BigEndian, ReadBytesExt};
-use num_enum::{IntoPrimitive, TryFromPrimitive};
-use serde_bytes::ByteBuf;
-use serde_cbor::Value;
-use serde_derive::{Deserialize, Serialize};
-use serde_indexed::{DeserializeIndexed, SerializeIndexed};
-use serde_repr::{Deserialize_repr, Serialize_repr};
-use tracing::debug;
-use tracing::warn;
-
-use ctap_types::cose::PublicKey;
-
-use crate::ops::webauthn::GetAssertionRequest;
-use crate::ops::webauthn::MakeCredentialRequest;
 use crate::pin::PinUvAuthProtocol;
 use crate::proto::ctap1::Ctap1Transport;
-use crate::transport::error::CtapError;
+use num_enum::{IntoPrimitive, TryFromPrimitive};
+use serde_bytes::ByteBuf;
+use serde_derive::{Deserialize, Serialize};
+use serde_repr::{Deserialize_repr, Serialize_repr};
+
+mod get_info;
+pub use get_info::Ctap2GetInfoResponse;
+mod bio_enrollment;
+pub use bio_enrollment::{
+    Ctap2BioEnrollmentFingerprintKind, Ctap2BioEnrollmentModality, Ctap2BioEnrollmentRequest,
+    Ctap2BioEnrollmentResponse, Ctap2BioEnrollmentTemplateId, Ctap2LastEnrollmentSampleStatus,
+};
+mod authenticator_config;
+pub use authenticator_config::{
+    Ctap2AuthenticatorConfigCommand, Ctap2AuthenticatorConfigParams,
+    Ctap2AuthenticatorConfigRequest,
+};
+mod client_pin;
+pub use client_pin::{
+    Ctap2AuthTokenPermissionRole, Ctap2ClientPinRequest, Ctap2ClientPinResponse,
+    Ctap2PinUvAuthProtocol,
+};
+mod make_credential;
+pub use make_credential::{
+    Ctap2MakeCredentialOptions, Ctap2MakeCredentialRequest, Ctap2MakeCredentialResponse,
+};
+mod get_assertion;
+pub use get_assertion::{
+    Ctap2AttestationStatement, Ctap2GetAssertionRequest, Ctap2GetAssertionResponse,
+    FidoU2fAttestationStmt,
+};
 
 // 32 (rpIdHash) + 1 (flags) + 4 (signCount) + 16 (aaguid
 const AUTHENTICATOR_DATA_PUBLIC_KEY_OFFSET: usize = 53;
@@ -132,33 +142,6 @@ pub struct Ctap2PublicKeyCredentialDescriptor {
     pub transports: Option<Vec<Ctap2Transport>>,
 }
 
-impl TryFrom<&Ctap2MakeCredentialResponse> for Ctap2PublicKeyCredentialDescriptor {
-    type Error = CtapError;
-    fn try_from(response: &Ctap2MakeCredentialResponse) -> Result<Self, Self::Error> {
-        if response.authenticator_data.len() < AUTHENTICATOR_DATA_PUBLIC_KEY_OFFSET + 2 {
-            warn!("Failed to parse credential ID: invalid authenticator data length");
-            return Err(CtapError::InvalidCredential);
-        }
-
-        let mut cursor = IOCursor::new(response.authenticator_data.as_ref());
-        cursor.set_position(AUTHENTICATOR_DATA_PUBLIC_KEY_OFFSET as u64);
-        let len = cursor.read_u16::<BigEndian>().unwrap() as usize;
-        let offset = AUTHENTICATOR_DATA_PUBLIC_KEY_OFFSET + 2;
-        if response.authenticator_data.len() < offset + len {
-            warn!("Failed to parse credential ID: not enough bytes");
-            return Err(CtapError::InvalidCredential);
-        }
-
-        let credential_id = response.authenticator_data[offset..offset + len].to_vec();
-        assert_eq!(len, credential_id.len());
-        Ok(Ctap2PublicKeyCredentialDescriptor {
-            r#type: Ctap2PublicKeyCredentialType::PublicKey,
-            id: ByteBuf::from(credential_id),
-            transports: None,
-        })
-    }
-}
-
 #[repr(i32)]
 #[derive(Debug, Clone, Copy, FromPrimitive, PartialEq, Serialize_repr, Deserialize_repr)]
 pub enum Ctap2COSEAlgorithmIdentifier {
@@ -197,281 +180,6 @@ impl Ctap2CredentialType {
     }
 }
 
-#[derive(Debug, Clone, Copy, Serialize)]
-pub struct Ctap2MakeCredentialOptions {
-    #[serde(rename = "rk")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub require_resident_key: Option<bool>,
-
-    #[serde(rename = "uv")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub deprecated_require_user_verification: Option<bool>,
-}
-
-impl Default for Ctap2MakeCredentialOptions {
-    fn default() -> Self {
-        Self {
-            require_resident_key: None,
-            deprecated_require_user_verification: None,
-        }
-    }
-}
-
-impl Ctap2MakeCredentialOptions {
-    pub fn skip_serializing(&self) -> bool {
-        self.require_resident_key.is_none() && self.deprecated_require_user_verification.is_none()
-    }
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Default)]
-pub struct Ctap2GetAssertionOptions {
-    #[serde(rename = "up")]
-    /// True for all requests; False for pre-flight only.
-    pub require_user_presence: bool,
-
-    #[serde(rename = "uv")]
-    #[serde(skip_serializing_if = "Self::skip_serializing_uv")]
-    pub require_user_verification: bool,
-}
-
-impl Ctap2GetAssertionOptions {
-    fn skip_serializing_uv(uv: &bool) -> bool {
-        !uv
-    }
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub struct PackedAttestationStmt {
-    #[serde(rename = "alg")]
-    pub algorithm: Ctap2COSEAlgorithmIdentifier,
-
-    #[serde(rename = "sig")]
-    pub signature: ByteBuf,
-
-    #[serde(rename = "x5c")]
-    pub certificates: Vec<ByteBuf>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub struct FidoU2fAttestationStmt {
-    #[serde(rename = "alg")]
-    pub algorithm: Ctap2COSEAlgorithmIdentifier,
-
-    #[serde(rename = "sig")]
-    pub signature: ByteBuf,
-
-    #[serde(rename = "x5c")]
-    pub certificates: Vec<ByteBuf>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub struct TpmAttestationStmt {
-    #[serde(rename = "ver")]
-    pub version: String,
-
-    #[serde(rename = "alg")]
-    pub algorithm: Ctap2COSEAlgorithmIdentifier,
-
-    #[serde(rename = "sig")]
-    pub signature: ByteBuf,
-
-    #[serde(rename = "x5c")]
-    pub certificates: Vec<ByteBuf>,
-
-    #[serde(rename = "certInfo")]
-    pub certificate_info: ByteBuf,
-
-    #[serde(rename = "pubArea")]
-    pub public_area: ByteBuf,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub struct AppleAnonymousAttestationStmt {
-    #[serde(rename = "x5c")]
-    pub certificates: Vec<ByteBuf>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(untagged)]
-pub enum Ctap2AttestationStatement {
-    PackedOrAndroid(PackedAttestationStmt),
-    Tpm(TpmAttestationStmt),
-    FidoU2F(FidoU2fAttestationStmt),
-    AppleAnonymous(AppleAnonymousAttestationStmt),
-    None(BTreeMap<Value, Value>),
-}
-
-// https://www.w3.org/TR/webauthn/#authenticatormakecredential
-#[derive(Debug, Clone, SerializeIndexed)]
-#[serde_indexed(offset = 1)]
-pub struct Ctap2MakeCredentialRequest {
-    /// clientDataHash (0x01)
-    pub hash: ByteBuf,
-
-    /// rp (0x02)
-    pub relying_party: Ctap2PublicKeyCredentialRpEntity,
-
-    /// user (0x03)
-    pub user: Ctap2PublicKeyCredentialUserEntity,
-
-    /// pubKeyCredParams (0x04)
-    pub algorithms: Vec<Ctap2CredentialType>,
-
-    /// excludeList (0x05)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub exclude: Option<Vec<Ctap2PublicKeyCredentialDescriptor>>,
-
-    /// extensions (0x06)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub extensions_cbor: Option<Vec<u8>>,
-
-    /// options (0x07)
-    #[serde(skip_serializing_if = "Self::skip_serializing_options")]
-    pub options: Option<Ctap2MakeCredentialOptions>,
-
-    /// pinUvAuthParam (0x08)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pin_auth_param: Option<ByteBuf>,
-
-    /// pinUvAuthProtocol (0x09)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pin_auth_proto: Option<u32>,
-
-    /// enterpriseAttestation (0x0A)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub enterprise_attestation: Option<u32>,
-}
-
-impl Ctap2MakeCredentialRequest {
-    pub fn skip_serializing_options(options: &Option<Ctap2MakeCredentialOptions>) -> bool {
-        options.map_or(true, |options| options.skip_serializing())
-    }
-}
-
-impl From<&MakeCredentialRequest> for Ctap2MakeCredentialRequest {
-    fn from(op: &MakeCredentialRequest) -> Ctap2MakeCredentialRequest {
-        Ctap2MakeCredentialRequest {
-            hash: ByteBuf::from(op.hash.clone()),
-            relying_party: op.relying_party.clone(),
-            user: op.user.clone(),
-            algorithms: op.algorithms.clone(),
-            exclude: op.exclude.clone(),
-            extensions_cbor: if op.extensions_cbor.is_empty() {
-                None
-            } else {
-                Some(op.extensions_cbor.clone())
-            },
-            options: Some(Ctap2MakeCredentialOptions {
-                require_resident_key: if op.require_resident_key {
-                    Some(true)
-                } else {
-                    None
-                },
-                deprecated_require_user_verification: None,
-            }),
-            pin_auth_param: None,
-            pin_auth_proto: None,
-            enterprise_attestation: None,
-        }
-    }
-}
-
-#[derive(Debug, Clone, DeserializeIndexed)]
-#[serde_indexed(offset = 1)]
-pub struct Ctap2MakeCredentialResponse {
-    pub format: String,
-    pub authenticator_data: ByteBuf,
-    pub attestation_statement: Ctap2AttestationStatement,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub enterprise_attestation: Option<bool>,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub large_blob_key: Option<ByteBuf>,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub unsigned_extension_output: Option<BTreeMap<Value, Value>>,
-}
-
-// https://www.w3.org/TR/webauthn/#op-get-assertion
-#[derive(Debug, Clone, SerializeIndexed)]
-#[serde_indexed(offset = 1)]
-pub struct Ctap2GetAssertionRequest {
-    /// rpId (0x01)
-    pub relying_party_id: String,
-
-    /// clientDataHash (0x02)
-    pub client_data_hash: ByteBuf,
-
-    /// allowList (0x03)
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub allow: Vec<Ctap2PublicKeyCredentialDescriptor>,
-
-    /// extensions (0x04)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub extensions_cbor: Option<Vec<u8>>,
-
-    /// options (0x05)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub options: Option<Ctap2GetAssertionOptions>,
-
-    /// pinUvAuthParam (0x06)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pin_auth_param: Option<ByteBuf>,
-
-    /// pinUvAuthProtocol (0x07)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pin_auth_proto: Option<u32>,
-}
-
-impl From<&GetAssertionRequest> for Ctap2GetAssertionRequest {
-    fn from(op: &GetAssertionRequest) -> Self {
-        Self {
-            relying_party_id: op.relying_party_id.clone(),
-            client_data_hash: ByteBuf::from(op.hash.clone()),
-            allow: op.allow.clone(),
-            extensions_cbor: op.extensions_cbor.clone(),
-            options: Some(Ctap2GetAssertionOptions {
-                require_user_presence: true,
-                require_user_verification: op.user_verification.is_required(),
-            }),
-            pin_auth_param: None,
-            pin_auth_proto: None,
-        }
-    }
-}
-
-#[derive(Debug, Clone, DeserializeIndexed)]
-#[serde_indexed(offset = 1)]
-pub struct Ctap2GetAssertionResponse {
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub credential_id: Option<Ctap2PublicKeyCredentialDescriptor>,
-    pub authenticator_data: ByteBuf,
-    pub signature: ByteBuf,
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub user: Option<Ctap2PublicKeyCredentialUserEntity>,
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub credentials_count: Option<u32>,
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub user_selected: Option<bool>,
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub large_blob_key: Option<ByteBuf>,
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub unsigned_extension_outputs: Option<BTreeMap<Value, Value>>,
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub enterprise_attestation: Option<bool>,
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub attestation_statement: Option<Ctap2AttestationStatement>,
-}
-
 pub trait Ctap2UserVerifiableRequest {
     fn ensure_uv_set(&mut self);
     fn calculate_and_set_uv_auth(
@@ -485,858 +193,10 @@ pub trait Ctap2UserVerifiableRequest {
     fn can_use_uv(&self, info: &Ctap2GetInfoResponse) -> bool;
 }
 
-impl Ctap2UserVerifiableRequest for Ctap2MakeCredentialRequest {
-    fn ensure_uv_set(&mut self) {
-        self.options = Some(Ctap2MakeCredentialOptions {
-            deprecated_require_user_verification: Some(true),
-            ..self
-                .options
-                .unwrap_or(Ctap2MakeCredentialOptions::default())
-        });
-    }
-
-    fn calculate_and_set_uv_auth(
-        &mut self,
-        uv_proto: &Box<dyn PinUvAuthProtocol>,
-        uv_auth_token: &[u8],
-    ) {
-        let uv_auth_param = uv_proto.authenticate(uv_auth_token, self.client_data_hash());
-        self.pin_auth_proto = Some(uv_proto.version() as u32);
-        self.pin_auth_param = Some(ByteBuf::from(uv_auth_param));
-    }
-
-    fn client_data_hash(&self) -> &[u8] {
-        self.hash.as_slice()
-    }
-
-    fn permissions(&self) -> Ctap2AuthTokenPermissionRole {
-        // GET_ASSERTION needed for pre-flight requests
-        return Ctap2AuthTokenPermissionRole::MAKE_CREDENTIAL
-            | Ctap2AuthTokenPermissionRole::GET_ASSERTION;
-    }
-
-    fn permissions_rpid(&self) -> Option<&str> {
-        Some(&self.relying_party.id)
-    }
-
-    fn can_use_uv(&self, _info: &Ctap2GetInfoResponse) -> bool {
-        true
-    }
-}
-
-impl Ctap2UserVerifiableRequest for Ctap2GetAssertionRequest {
-    fn ensure_uv_set(&mut self) {
-        self.options = Some(Ctap2GetAssertionOptions {
-            require_user_verification: true,
-            ..self.options.unwrap_or(Ctap2GetAssertionOptions::default())
-        });
-    }
-
-    fn calculate_and_set_uv_auth(
-        &mut self,
-        uv_proto: &Box<dyn PinUvAuthProtocol>,
-        uv_auth_token: &[u8],
-    ) {
-        let uv_auth_param = uv_proto.authenticate(uv_auth_token, self.client_data_hash());
-        self.pin_auth_proto = Some(uv_proto.version() as u32);
-        self.pin_auth_param = Some(ByteBuf::from(uv_auth_param));
-    }
-
-    fn client_data_hash(&self) -> &[u8] {
-        self.client_data_hash.as_slice()
-    }
-
-    fn permissions(&self) -> Ctap2AuthTokenPermissionRole {
-        return Ctap2AuthTokenPermissionRole::GET_ASSERTION;
-    }
-
-    fn permissions_rpid(&self) -> Option<&str> {
-        Some(&self.relying_party_id)
-    }
-
-    fn can_use_uv(&self, _info: &Ctap2GetInfoResponse) -> bool {
-        true
-    }
-}
-
-#[derive(Debug, Clone, DeserializeIndexed)]
-#[serde_indexed(offset = 1)]
-pub struct Ctap2GetInfoResponse {
-    /// versions (0x01)
-    pub versions: Vec<String>,
-
-    /// extensions (0x02)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub extensions: Option<Vec<String>>,
-
-    /// aaguid (0x03)
-    pub aaguid: ByteBuf,
-
-    /// options (0x04)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub options: Option<HashMap<String, bool>>,
-
-    /// maxMsgSize (0x05)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_msg_size: Option<u32>,
-
-    /// pinUvAuthProtocols (0x06)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pin_auth_protos: Option<Vec<u32>>,
-
-    /// maxCredentialCountInList (0x07)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_credential_count: Option<u32>,
-
-    /// maxCredentialIdLength (0x08)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_credential_id_length: Option<u32>,
-
-    /// transports (0x09)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub transports: Option<Vec<String>>,
-
-    /// algorithms (0x0A)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub algorithms: Option<Vec<Ctap2CredentialType>>,
-
-    /// maxSerializedLargeBlobArray (0x0B)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_blob_array: Option<u32>,
-
-    /// forcePINChange (0x0C)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub force_pin_change: Option<bool>,
-
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub min_pin_length: Option<u32>,
-
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub firmware_version: Option<u32>,
-
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_cred_blob_length: Option<u32>,
-
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_rpids_for_setminpinlength: Option<u32>,
-
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub preferred_platform_uv_attempts: Option<u32>,
-
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub uv_modality: Option<u32>,
-
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub certifications: Option<HashMap<String, u32>>,
-
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub remaining_discoverable_creds: Option<u32>,
-
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub vendor_proto_config_cmds: Option<Vec<u32>>,
-}
-
 #[derive(Debug, Clone, Copy)]
 pub enum Ctap2UserVerificationOperation {
     GetPinUvAuthTokenUsingUvWithPermissions,
     GetPinUvAuthTokenUsingPinWithPermissions,
     GetPinToken,
     None,
-}
-
-impl Ctap2GetInfoResponse {
-    pub fn option_enabled(&self, name: &str) -> bool {
-        if self.options.is_none() {
-            return false;
-        }
-        let options = self.options.as_ref().unwrap();
-        options.get(name) == Some(&true)
-    }
-
-    pub fn supports_fido_2_1(&self) -> bool {
-        self.versions.iter().any(|v| v == "FIDO_2_1")
-    }
-
-    /// Implements check for "Protected by some form of User Verification":
-    ///   Either or both clientPin or built-in user verification methods are supported and enabled.
-    ///   I.e., in the authenticatorGetInfo response the pinUvAuthToken option ID is present and set to true,
-    ///   and either clientPin option ID is present and set to true or uv option ID is present and set to true or both.
-    pub fn is_uv_protected(&self) -> bool {
-        self.option_enabled("uv") || // Deprecated no-op UV
-            self.option_enabled("clientPin") ||
-            (self.option_enabled("pinUvAuthToken") && self.option_enabled("uv"))
-    }
-
-    pub fn uv_operation(&self, uv_blocked: bool) -> Option<Ctap2UserVerificationOperation> {
-        if self.option_enabled("uv") && !uv_blocked {
-            if self.option_enabled("pinUvAuthToken") {
-                debug!("getPinUvAuthTokenUsingUvWithPermissions");
-                return Some(
-                    Ctap2UserVerificationOperation::GetPinUvAuthTokenUsingUvWithPermissions,
-                );
-            } else {
-                debug!("Deprecated FIDO 2.0 behaviour: populating 'uv' flag");
-                return Some(Ctap2UserVerificationOperation::None);
-            }
-        } else {
-            // !uv
-            if self.option_enabled("pinUvAuthToken") {
-                assert!(self.option_enabled("clientPin"));
-                debug!("getPinUvAuthTokenUsingPinWithPermissions");
-                return Some(
-                    Ctap2UserVerificationOperation::GetPinUvAuthTokenUsingPinWithPermissions,
-                );
-            } else if self.option_enabled("clientPin") {
-                // !pinUvAuthToken
-                debug!("getPinToken");
-                return Some(Ctap2UserVerificationOperation::GetPinToken);
-            } else {
-                debug!("No UV and no PIN (e.g. maybe UV was blocked and no PIN available)");
-                return None;
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone, SerializeIndexed)]
-#[serde_indexed(offset = 1)]
-pub struct Ctap2ClientPinRequest {
-    ///pinUvAuthProtocol (0x01)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub protocol: Option<Ctap2PinUvAuthProtocol>,
-
-    /// subCommand (0x02)
-    pub command: Ctap2PinUvAuthProtocolCommand,
-
-    /// keyAgreement (0x03)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub key_agreement: Option<PublicKey>,
-
-    /// pinUvAuthParam (0x04):
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub uv_auth_param: Option<ByteBuf>,
-
-    /// newPinEnc (0x05)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub new_pin_encrypted: Option<ByteBuf>,
-
-    /// pinHashEnc (0x06)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pin_hash_encrypted: Option<ByteBuf>,
-
-    #[serde(skip_serializing)]
-    pub unused_07: (),
-
-    #[serde(skip_serializing)]
-    pub unused_08: (),
-
-    /// permissions (0x09)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub permissions: Option<u32>,
-
-    /// permissions RPID (0x10)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub permissions_rpid: Option<String>,
-}
-
-impl Ctap2ClientPinRequest {
-    pub fn new_get_key_agreement(protocol: Ctap2PinUvAuthProtocol) -> Self {
-        Self {
-            protocol: Some(protocol),
-            command: Ctap2PinUvAuthProtocolCommand::GetKeyAgreement,
-            key_agreement: None,
-            uv_auth_param: None,
-            new_pin_encrypted: None,
-            pin_hash_encrypted: None,
-            unused_07: (),
-            unused_08: (),
-            permissions: None,
-            permissions_rpid: None,
-        }
-    }
-
-    pub fn new_get_pin_token(
-        protocol: Ctap2PinUvAuthProtocol,
-        public_key: PublicKey,
-        pin_hash_enc: &[u8],
-    ) -> Self {
-        Self {
-            protocol: Some(protocol),
-            command: Ctap2PinUvAuthProtocolCommand::GetPinToken,
-            key_agreement: Some(public_key),
-            uv_auth_param: None,
-            new_pin_encrypted: None,
-            pin_hash_encrypted: Some(ByteBuf::from(pin_hash_enc)),
-            unused_07: (),
-            unused_08: (),
-            permissions: None,
-            permissions_rpid: None,
-        }
-    }
-
-    pub fn new_get_pin_retries() -> Self {
-        Self {
-            protocol: None,
-            command: Ctap2PinUvAuthProtocolCommand::GetPinRetries,
-            key_agreement: None,
-            uv_auth_param: None,
-            new_pin_encrypted: None,
-            pin_hash_encrypted: None,
-            unused_07: (),
-            unused_08: (),
-            permissions: None,
-            permissions_rpid: None,
-        }
-    }
-
-    pub fn new_get_pin_token_with_perm(
-        protocol: Ctap2PinUvAuthProtocol,
-        public_key: PublicKey,
-        pin_hash_enc: &[u8],
-        permissions: Ctap2AuthTokenPermissionRole,
-        permissions_rpid: Option<&str>,
-    ) -> Self {
-        Self {
-            protocol: Some(protocol),
-            command: Ctap2PinUvAuthProtocolCommand::GetPinUvAuthTokenUsingPinWithPermissions,
-            key_agreement: Some(public_key),
-            uv_auth_param: None,
-            new_pin_encrypted: None,
-            pin_hash_encrypted: Some(ByteBuf::from(pin_hash_enc)),
-            unused_07: (),
-            unused_08: (),
-            permissions: Some(permissions.bits()),
-            permissions_rpid: permissions_rpid.map(str::to_owned),
-        }
-    }
-
-    pub fn new_get_uv_token_with_perm(
-        protocol: Ctap2PinUvAuthProtocol,
-        public_key: PublicKey,
-        permissions: Ctap2AuthTokenPermissionRole,
-        permissions_rpid: Option<&str>,
-    ) -> Self {
-        Self {
-            protocol: Some(protocol),
-            command: Ctap2PinUvAuthProtocolCommand::GetPinUvAuthTokenUsingUvWithPermissions,
-            key_agreement: Some(public_key),
-            uv_auth_param: None,
-            new_pin_encrypted: None,
-            pin_hash_encrypted: None,
-            unused_07: (),
-            unused_08: (),
-            permissions: Some(permissions.bits()),
-            permissions_rpid: permissions_rpid.map(str::to_owned),
-        }
-    }
-
-    pub fn new_change_pin(
-        protocol: Ctap2PinUvAuthProtocol,
-        new_pin_enc: &[u8],
-        curr_pin_enc: &[u8],
-        public_key: PublicKey,
-        uv_auth_param: &[u8],
-    ) -> Self {
-        Self {
-            protocol: Some(protocol),
-            command: Ctap2PinUvAuthProtocolCommand::ChangePin,
-            key_agreement: Some(public_key),
-            uv_auth_param: Some(ByteBuf::from(uv_auth_param)),
-            new_pin_encrypted: Some(ByteBuf::from(new_pin_enc)),
-            pin_hash_encrypted: Some(ByteBuf::from(curr_pin_enc)),
-            unused_07: (),
-            unused_08: (),
-            permissions: None,
-            permissions_rpid: None,
-        }
-    }
-
-    pub fn new_set_pin(
-        protocol: Ctap2PinUvAuthProtocol,
-        new_pin_enc: &[u8],
-        public_key: PublicKey,
-        uv_auth_param: &[u8],
-    ) -> Self {
-        Self {
-            protocol: Some(protocol),
-            command: Ctap2PinUvAuthProtocolCommand::SetPin,
-            key_agreement: Some(public_key),
-            uv_auth_param: Some(ByteBuf::from(uv_auth_param)),
-            new_pin_encrypted: Some(ByteBuf::from(new_pin_enc)),
-            pin_hash_encrypted: None,
-            unused_07: (),
-            unused_08: (),
-            permissions: None,
-            permissions_rpid: None,
-        }
-    }
-}
-
-bitflags! {
-    #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-    pub struct Ctap2AuthTokenPermissionRole: u32 {
-        const MAKE_CREDENTIAL = 0x01;
-        const GET_ASSERTION = 0x02;
-        const CREDENTIAL_MANAGEMENT = 0x04;
-        const BIO_ENROLLMENT = 0x08;
-        const LARGE_BLOB_WRITE = 0x10;
-        const AUTHENTICATOR_CONFIGURATION = 0x20;
-    }
-}
-
-#[repr(u32)]
-#[derive(Debug, Clone, Copy, FromPrimitive, PartialEq, Eq, Serialize_repr, Deserialize_repr)]
-pub enum Ctap2PinUvAuthProtocol {
-    One = 1,
-    Two = 2,
-}
-
-#[repr(u32)]
-#[derive(Debug, Clone, FromPrimitive, PartialEq, Serialize_repr, Deserialize_repr)]
-pub enum Ctap2PinUvAuthProtocolCommand {
-    GetPinRetries = 0x01,
-    GetKeyAgreement = 0x02,
-    SetPin = 0x03,
-    ChangePin = 0x04,
-    GetPinToken = 0x05,
-    GetPinUvAuthTokenUsingUvWithPermissions = 0x06,
-    GetUvRetries = 0x07,
-    GetPinUvAuthTokenUsingPinWithPermissions = 0x09,
-}
-
-#[derive(Debug, Clone, Default, DeserializeIndexed)]
-#[serde_indexed(offset = 1)]
-pub struct Ctap2ClientPinResponse {
-    /// keyAgreement (0x01)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub key_agreement: Option<PublicKey>,
-
-    /// pinUvAuthToken (0x02)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pin_uv_auth_token: Option<ByteBuf>,
-
-    /// pinRetries (0x03)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pin_retries: Option<u32>,
-
-    /// powerCycleState (0x04)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub power_cycle_state: Option<bool>,
-
-    /// uvRetries (0x05)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub uv_retries: Option<u32>,
-}
-
-#[derive(Debug, Clone, SerializeIndexed)]
-#[serde_indexed(offset = 1)]
-pub struct Ctap2AuthenticatorConfigRequest {
-    // subCommand (0x01)
-    pub subcommand: Ctap2AuthenticatorConfigCommand,
-
-    // subCommandParams (0x02)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub subcommand_params: Option<Ctap2AuthenticatorConfigParams>,
-
-    ///pinUvAuthProtocol (0x03)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub protocol: Option<Ctap2PinUvAuthProtocol>,
-
-    /// pinUvAuthParam (0x04):
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub uv_auth_param: Option<ByteBuf>,
-}
-
-impl Ctap2AuthenticatorConfigRequest {
-    pub(crate) fn new_toggle_always_uv() -> Self {
-        Ctap2AuthenticatorConfigRequest {
-            subcommand: Ctap2AuthenticatorConfigCommand::ToggleAlwaysUv,
-            subcommand_params: None,
-            protocol: None,      // Will be filled out later by user_verification()
-            uv_auth_param: None, // Will be filled out later by user_verification()
-        }
-    }
-
-    pub(crate) fn new_enable_enterprise_attestation() -> Self {
-        Ctap2AuthenticatorConfigRequest {
-            subcommand: Ctap2AuthenticatorConfigCommand::EnableEnterpriseAttestation,
-            subcommand_params: None,
-            protocol: None,      // Will be filled out later by user_verification()
-            uv_auth_param: None, // Will be filled out later by user_verification()
-        }
-    }
-
-    pub(crate) fn new_force_change_pin(force_change_pin: bool) -> Self {
-        let subcommand_params =
-            Ctap2AuthenticatorConfigParams::SetMinPINLength(Ctap2SetMinPINLengthParams {
-                new_min_pin_length: None,
-                min_pin_length_rpids: None,
-                force_change_pin: Some(force_change_pin),
-            });
-        Ctap2AuthenticatorConfigRequest {
-            subcommand: Ctap2AuthenticatorConfigCommand::SetMinPINLength,
-            subcommand_params: Some(subcommand_params),
-            protocol: None,      // Will be filled out later by user_verification()
-            uv_auth_param: None, // Will be filled out later by user_verification()
-        }
-    }
-
-    pub(crate) fn new_set_min_pin_length(new_pin_length: u64) -> Self {
-        let subcommand_params =
-            Ctap2AuthenticatorConfigParams::SetMinPINLength(Ctap2SetMinPINLengthParams {
-                new_min_pin_length: Some(new_pin_length),
-                min_pin_length_rpids: None,
-                force_change_pin: None,
-            });
-        Ctap2AuthenticatorConfigRequest {
-            subcommand: Ctap2AuthenticatorConfigCommand::SetMinPINLength,
-            subcommand_params: Some(subcommand_params),
-            protocol: None,      // Will be filled out later by user_verification()
-            uv_auth_param: None, // Will be filled out later by user_verification()
-        }
-    }
-
-    pub(crate) fn new_set_min_pin_length_rpids(rpids: Vec<String>) -> Self {
-        let subcommand_params =
-            Ctap2AuthenticatorConfigParams::SetMinPINLengthRPIDs(Ctap2SetMinPINLengthParams {
-                new_min_pin_length: None,
-                min_pin_length_rpids: Some(rpids),
-                force_change_pin: None,
-            });
-        Ctap2AuthenticatorConfigRequest {
-            subcommand: Ctap2AuthenticatorConfigCommand::SetMinPINLength,
-            subcommand_params: Some(subcommand_params),
-            protocol: None,      // Will be filled out later by user_verification()
-            uv_auth_param: None, // Will be filled out later by user_verification()
-        }
-    }
-}
-
-#[repr(u32)]
-#[derive(Debug, Copy, Clone, FromPrimitive, PartialEq, Serialize_repr, Deserialize_repr)]
-pub enum Ctap2AuthenticatorConfigCommand {
-    EnableEnterpriseAttestation = 0x01,
-    ToggleAlwaysUv = 0x02,
-    SetMinPINLength = 0x03,
-    VendorPrototype = 0xFF,
-}
-
-#[derive(Debug, Clone, Serialize)]
-#[serde(untagged)]
-pub enum Ctap2AuthenticatorConfigParams {
-    SetMinPINLength(Ctap2SetMinPINLengthParams),
-    SetMinPINLengthRPIDs(Ctap2SetMinPINLengthParams),
-}
-
-#[derive(Debug, Clone, SerializeIndexed)]
-#[serde_indexed(offset = 1)]
-pub struct Ctap2SetMinPINLengthParams {
-    // newMinPINLength (0x01)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub new_min_pin_length: Option<u64>,
-
-    // minPinLengthRPIDs (0x02)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub min_pin_length_rpids: Option<Vec<String>>,
-
-    // forceChangePin (0x03)
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub force_change_pin: Option<bool>,
-}
-
-#[derive(Debug, Clone, SerializeIndexed)]
-#[serde_indexed(offset = 1)]
-pub struct Ctap2BioEnrollmentRequest {
-    // modality (0x01) 	Unsigned Integer 	Optional 	The user verification modality being requested
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub modality: Option<Ctap2BioEnrollmentModality>,
-
-    // subCommand (0x02) 	Unsigned Integer 	Optional 	The authenticator user verification sub command currently being requested
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub subcommand: Option<Ctap2BioEnrollmentSubcommand>,
-
-    // subCommandParams (0x03) 	CBOR Map 	Optional 	Map of subCommands parameters. This parameter MAY be omitted when the subCommand does not take any arguments.
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub subcommand_params: Option<Ctap2BioEnrollmentParams>,
-
-    // pinUvAuthProtocol (0x04) 	Unsigned Integer 	Optional 	PIN/UV protocol version chosen by the platform.
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub protocol: Option<Ctap2PinUvAuthProtocol>,
-
-    // pinUvAuthParam (0x05) 	Byte String 	Optional 	First 16 bytes of HMAC-SHA-256 of contents using pinUvAuthToken.
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub uv_auth_param: Option<ByteBuf>,
-
-    // getModality (0x06) 	Boolean 	Optional 	Get the user verification type modality. This MUST be set to true.
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub get_modality: Option<bool>,
-}
-
-#[repr(u32)]
-#[derive(Debug, Copy, Clone, FromPrimitive, PartialEq, Serialize_repr, Deserialize_repr)]
-pub enum Ctap2BioEnrollmentSubcommand {
-    EnrollBegin = 0x01,
-    EnrollCaptureNextSample = 0x02,
-    CancelCurrentEnrollment = 0x03,
-    EnumerateEnrollments = 0x04,
-    SetFriendlyName = 0x05,
-    RemoveEnrollment = 0x06,
-    GetFingerprintSensorInfo = 0x07,
-}
-
-#[derive(Debug, Clone, SerializeIndexed)]
-#[serde_indexed(offset = 1)]
-pub struct Ctap2BioEnrollmentParams {
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    template_id: Option<ByteBuf>,
-
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    template_friendly_name: Option<String>,
-
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    timeout_milliseconds: Option<u64>,
-}
-
-#[derive(Debug, Default, Clone, DeserializeIndexed)]
-#[serde_indexed(offset = 1)]
-pub struct Ctap2BioEnrollmentResponse {
-    // modality (0x01) 	Unsigned Integer 	Optional 	The user verification modality.
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub modality: Option<Ctap2BioEnrollmentModality>,
-
-    // fingerprintKind (0x02) 	Unsigned Integer 	Optional 	Indicates the type of fingerprint sensor. For touch type sensor, its value is 1. For swipe type sensor its value is 2.
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub fingerprint_kind: Option<Ctap2BioEnrollmentFingerprintKind>,
-
-    // maxCaptureSamplesRequiredForEnroll (0x03) 	Unsigned Integer 	Optional 	Indicates the maximum good samples required for enrollment.
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_capture_samples_required_for_enroll: Option<u64>,
-
-    // templateId (0x04) 	Byte String 	Optional 	Template Identifier.
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub template_id: Option<ByteBuf>,
-
-    // lastEnrollSampleStatus (0x05) 	Unsigned Integer 	Optional 	Last enrollment sample status.
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub last_enroll_sample_status: Option<Ctap2LastEnrollmentSampleStatus>,
-
-    // remainingSamples (0x06) 	Unsigned Integer 	Optional 	Number of more sample required for enrollment to complete
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub remaining_samples: Option<u64>,
-
-    // templateInfos (0x07) 	CBOR ARRAY 	Optional 	Array of templateInfo’s
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub template_infos: Option<Vec<Ctap2BioEnrollmentTemplateId>>,
-
-    // maxTemplateFriendlyName (0x08) 	Unsigned Integer 	Optional 	Indicates the maximum number of bytes the authenticator will accept as a templateFriendlyName.
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_template_friendly_name: Option<u64>,
-}
-
-#[derive(Debug, Clone, DeserializeIndexed)]
-#[serde_indexed(offset = 1)]
-pub struct Ctap2BioEnrollmentTemplateId {
-    // templateId (0x01) 	Byte String 	Required 	Template Identifier.
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub template_id: Option<ByteBuf>,
-
-    // templateFriendlyName (0x02) 	String 	Optional 	Template Friendly Name.
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub template_friendly_name: Option<String>,
-}
-
-#[repr(u64)]
-#[derive(Debug, Clone, Copy, FromPrimitive, PartialEq, Serialize_repr, Deserialize_repr)]
-pub enum Ctap2LastEnrollmentSampleStatus {
-    Ctap2EnrollFeedbackFpGood = 0x00,        // Good fingerprint capture.
-    Ctap2EnrollFeedbackFpTooHigh = 0x01,     // Fingerprint was too high.
-    Ctap2EnrollFeedbackFpTooLow = 0x02,      // Fingerprint was too low.
-    Ctap2EnrollFeedbackFpTooLeft = 0x03,     // Fingerprint was too left.
-    Ctap2EnrollFeedbackFpTooRight = 0x04,    // Fingerprint was too right.
-    Ctap2EnrollFeedbackFpTooFast = 0x05,     // Fingerprint was too fast.
-    Ctap2EnrollFeedbackFpTooSlow = 0x06,     // Fingerprint was too slow.
-    Ctap2EnrollFeedbackFpPoorQuality = 0x07, // Fingerprint was of poor quality.
-    Ctap2EnrollFeedbackFpTooSkewed = 0x08,   // Fingerprint was too skewed.
-    Ctap2EnrollFeedbackFpTooShort = 0x09,    // Fingerprint was too short.
-    Ctap2EnrollFeedbackFpMergeFailure = 0x0A, // Merge failure of the capture.
-    Ctap2EnrollFeedbackFpExists = 0x0B,      // Fingerprint already exists.
-    Unused = 0x0C,                           // (this error number is available)
-    Ctap2EnrollFeedbackNoUserActivity = 0x0D, // User did not touch/swipe the authenticator.
-    Ctap2EnrollFeedbackNoUserPresenceTransition = 0x0E, // User did not lift the finger off the sensor.
-}
-
-#[repr(u64)]
-#[derive(Debug, Clone, FromPrimitive, PartialEq, Serialize_repr, Deserialize_repr)]
-pub enum Ctap2BioEnrollmentModality {
-    Fingerprint = 0x01, // Fingerprint was too high.
-}
-
-#[repr(u64)]
-#[derive(Debug, Clone, FromPrimitive, PartialEq, Serialize_repr, Deserialize_repr)]
-pub enum Ctap2BioEnrollmentFingerprintKind {
-    Touch = 0x01,
-    Swipe = 0x02,
-}
-
-impl Ctap2BioEnrollmentRequest {
-    pub fn new_get_modality() -> Self {
-        Ctap2BioEnrollmentRequest {
-            modality: None,
-            subcommand: None,
-            subcommand_params: None,
-            protocol: None,      // Get's filled in later
-            uv_auth_param: None, // Get's filled in later
-            get_modality: Some(true),
-        }
-    }
-
-    pub fn new_fingerprint_sensor_info() -> Self {
-        Ctap2BioEnrollmentRequest {
-            modality: Some(Ctap2BioEnrollmentModality::Fingerprint),
-            subcommand: Some(Ctap2BioEnrollmentSubcommand::GetFingerprintSensorInfo),
-            subcommand_params: None,
-            protocol: None,      // Get's filled in later
-            uv_auth_param: None, // Get's filled in later
-            get_modality: None,
-        }
-    }
-
-    pub fn new_enumerate_enrollments() -> Self {
-        Ctap2BioEnrollmentRequest {
-            modality: Some(Ctap2BioEnrollmentModality::Fingerprint),
-            subcommand: Some(Ctap2BioEnrollmentSubcommand::EnumerateEnrollments),
-            subcommand_params: None,
-            protocol: None,      // Get's filled in later
-            uv_auth_param: None, // Get's filled in later
-            get_modality: None,
-        }
-    }
-
-    pub fn new_remove_enrollment(template_id: &[u8]) -> Self {
-        Ctap2BioEnrollmentRequest {
-            modality: Some(Ctap2BioEnrollmentModality::Fingerprint),
-            subcommand: Some(Ctap2BioEnrollmentSubcommand::RemoveEnrollment),
-            subcommand_params: Some(Ctap2BioEnrollmentParams {
-                template_id: Some(ByteBuf::from(template_id)),
-                template_friendly_name: None,
-                timeout_milliseconds: None,
-            }),
-            protocol: None,      // Get's filled in later
-            uv_auth_param: None, // Get's filled in later
-            get_modality: None,
-        }
-    }
-
-    pub fn new_rename_enrollment(template_id: &[u8], template_friendly_name: &str) -> Self {
-        Ctap2BioEnrollmentRequest {
-            modality: Some(Ctap2BioEnrollmentModality::Fingerprint),
-            subcommand: Some(Ctap2BioEnrollmentSubcommand::SetFriendlyName),
-            subcommand_params: Some(Ctap2BioEnrollmentParams {
-                template_id: Some(ByteBuf::from(template_id)),
-                template_friendly_name: Some(template_friendly_name.to_string()),
-                timeout_milliseconds: None,
-            }),
-            protocol: None,      // Get's filled in later
-            uv_auth_param: None, // Get's filled in later
-            get_modality: None,
-        }
-    }
-
-    pub fn new_start_new_enrollment(enrollment_timeout: Option<Duration>) -> Self {
-        let subcommand_params = if let Some(time) = enrollment_timeout {
-            Some(Ctap2BioEnrollmentParams {
-                template_id: None,
-                template_friendly_name: None,
-                timeout_milliseconds: Some(time.as_millis() as u64),
-            })
-        } else {
-            None
-        };
-
-        Ctap2BioEnrollmentRequest {
-            modality: Some(Ctap2BioEnrollmentModality::Fingerprint),
-            subcommand: Some(Ctap2BioEnrollmentSubcommand::EnrollBegin),
-            subcommand_params,
-            protocol: None,      // Get's filled in later
-            uv_auth_param: None, // Get's filled in later
-            get_modality: None,
-        }
-    }
-
-    pub fn new_next_enrollment(template_id: &[u8], enrollment_timeout: Option<Duration>) -> Self {
-        let subcommand_params = Some(Ctap2BioEnrollmentParams {
-            template_id: Some(ByteBuf::from(template_id)),
-            template_friendly_name: None,
-            timeout_milliseconds: enrollment_timeout.map(|x| x.as_millis() as u64),
-        });
-
-        Ctap2BioEnrollmentRequest {
-            modality: Some(Ctap2BioEnrollmentModality::Fingerprint),
-            subcommand: Some(Ctap2BioEnrollmentSubcommand::EnrollCaptureNextSample),
-            subcommand_params,
-            protocol: None,
-            uv_auth_param: None,
-            get_modality: None,
-        }
-    }
-
-    pub fn new_cancel_current_enrollment() -> Self {
-        Ctap2BioEnrollmentRequest {
-            modality: Some(Ctap2BioEnrollmentModality::Fingerprint),
-            subcommand: Some(Ctap2BioEnrollmentSubcommand::CancelCurrentEnrollment),
-            subcommand_params: None,
-            protocol: None,
-            uv_auth_param: None,
-            get_modality: None,
-        }
-    }
 }
